requirements.txt
# LEGS-SLAM pip dependencies
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# AI/ML dependencies
onnxruntime-gpu==1.20.1
einops==0.8.1
transformers==4.52.4
timm==1.0.15
triton==2.1.0
accelerate==0.25.0
diffusers==0.24.0
ninja==1.11.1
git+https://github.com/openai/CLIP.git
# Computer Vision
opencv-python==4.11.0.86
scikit-image==0.25.2
scikit-learn==1.7.0
pillow==11.1.0
matplotlib==3.10.0
imageio==2.37.0

# Data processing
numpy==1.26.4
scipy==1.15.2
pandas==2.2.3
pyyaml==6.0.2
omegaconf==2.3.0

# Utilities
tqdm==4.65.2
requests==2.28.2
rerun-sdk==0.22.0
rich==13.4.2
safetensors==0.5.3
tokenizers==0.21.1
protobuf==5.29.3

# OpenMMLab ecosystem
mmcls==0.25.0
mmcv-full==1.7.2
mmengine==0.10.6
mmsegmentation==0.30.0
openmim==0.3.9

# Cloud storage
# oss2==2.17.0
# opendatalab==0.0.10
# openxlab==0.1.2

# Other utilities
addict==2.4.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.1
cycler==0.12.1
filelock==3.14.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.2.0
ftfy==6.3.1
hf-xet==1.1.3
huggingface-hub==0.32.4
humanfriendly==10.0
jinja2==3.1.5
jmespath==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
lazy-loader==0.4
lightning-utilities==0.14.3
markdown==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
model-index==0.1.11
ordered-set==4.1.0
packaging==25.0
platformdirs==4.3.6
plyfile==1.1
prettytable==3.16.0
pyarrow==19.0.1
pycryptodome==3.21.0
pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2023.4
regex==2024.11.6
setuptools==60.2.0
six==1.17.0
sympy==1.13.1
tabulate==0.9.0
termcolor==2.5.0
threadpoolctl==3.6.0
tifffile==2025.5.10
torchmetrics==1.7.1
tomli==2.2.1
tzdata==2025.1
urllib3==1.26.20
wcwidth==0.2.13
yapf==0.43.0

# NVIDIA CUDA packages
# nvidia-cublas-cu11==11.11.3.6
# nvidia-cuda-cupti-cu11==11.8.87
# nvidia-cuda-nvrtc-cu11==11.8.89
# nvidia-cuda-runtime-cu11==11.8.89
# nvidia-cudnn-cu11==9.1.0.70
# nvidia-cufft-cu11==10.9.0.58
# nvidia-curand-cu11==10.3.0.86
# nvidia-cusolver-cu11==11.4.1.48
# nvidia-cusparse-cu11==11.7.5.86
# nvidia-nccl-cu11==2.21.5
# nvidia-nvtx-cu11==11.8.86 